Filename aux reserved on Windows, renamed to misc
