+++ src/misc.h
